added check for theme

+++ src/wdio/services/LightHouseService.js
@@ -1,19 +1,24 @@
 const fs = require('fs');
 const launchChromeAndRunLighthouse = require('../../../scripts/lighthouse/lighthouse');
 const { addReportData, generateReport } = require('../../../scripts/lighthouse/reportGenerator');
-const { terraViewports } = require('../../../config/wdio/services.default-config');
 
 const rootDir = 'performance_reports';
 const htmlRootDir = 'performance_reports//html';
+const consolidateHtmlReport = `${rootDir}//performance_report.html}`;
+
+/* Use to override default theme for theme visual regression tests. */
+const theme = process.env.THEME;
 
 export default class LightHouseService {
   before() {
-    if (fs.existsSync(rootDir)) {
+    // clears the previous report only for default theme. To prevent performance test from running for multiple themes.
+    if (fs.existsSync(rootDir) && theme === undefined) {
       const files = fs.readdirSync(htmlRootDir);
       files.forEach((filename) => {
         fs.unlinkSync(`${htmlRootDir}//${filename}`);
       });
-      fs.rmdirSync(htmlRootDir);
+      if (fs.existsSync(htmlRootDir)) fs.rmdirSync(htmlRootDir);
+      if (fs.existsSync(consolidateHtmlReport)) fs.unlinkSync(consolidateHtmlReport);
       fs.rmdirSync(rootDir);
     }
   }
@@ -38,6 +43,6 @@ export default class LightHouseService {
   }
 
   onComplete() {
-    generateReport();
+    generateReport(consolidateHtmlReport);
   }
 };
